backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-multipart==0.0.6
aiohttp==3.9.1
aiosqlite==0.19.0
pydantic==2.5.0
pydantic-settings==2.1.0
APScheduler==3.10.4
PyYAML==6.0.1
python-dateutil==2.8.2
websockets==12.0
